+++ libs/preprocessor/if.c
@@ -20,7 +20,7 @@
 #include "context_var.h"
 #include "file.h"
 #include "preprocessor.h"
-#include "preprocessor_error.h"
+#include "error.h"
 #include "preprocessor_utils.h"
 #include <limits.h>
 #include <math.h>
