+++ libs/preprocessor/preprocessor_utils.c
@@ -18,7 +18,7 @@
 #include "constants.h"
 #include "context_var.h"
 #include "file.h"
-#include "preprocessor_error.h"
+#include "error.h"
 #include "utf8.h"
 #include <limits.h>
 #include <math.h>
