+++ tests/executable/multifile/no_guard/include/test_1/test1.c
@@ -2,6 +2,7 @@
 #define B 3 
 #define C 7 
 #define D 1 
+
 void main()
 {
   int h = h1() + A;
